requirements.txt
# Phase 8: Production Dependencies for RAG-A2A-MCP System
# Core dependencies for containerized deployment

# Google ADK and AI APIs
google-adk>=1.0.0
openai>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# HTTP Client & Async
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.25.0
aiofiles>=23.0.0

# Security
cryptography>=41.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Monitoring & Logging
prometheus-client>=0.19.0
psutil>=5.9.0
structlog>=23.0.0

# Environment management
python-dotenv>=1.0.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Optional: Enhanced features (uncomment if needed)
# chromadb>=0.4.0
# langchain>=0.1.0
# litellm>=1.0.0
# fastmcp>=0.1.0
